Create Hash
We generate a secure hash function of the string "python_program", then detect if the code is changed. Hash functions produce a fix-sized value from variable length data. sha256 produces 256 bit numbers. One thing that we can do with this number is to see if the data has been modified or not.
One point here is that hash functions need a byte array. They don't work with strings. Because of that we've got python_program.encode() on line 8. The hexdigest method produces a hexadecimal representation of the secure hash. 
from line 11 on, we add some more python code to our data and generate the modified hash. When we run the program, the new hash is very different to the original one.